Added support for Single Store and added button to force feed generation. [sc-6419] [sc-6420]
