+++ Controller/Adminhtml/Feed/Generate.php
@@ -1,0 +1,79 @@
+<?php
+
+namespace Salesfire\Salesfire\Controller\Adminhtml\Feed;
+
+use Magento\Backend\App\Action;
+use Magento\Backend\App\Action\Context;
+use Magento\Framework\Controller\Result\JsonFactory;
+use Magento\Framework\View\Result\PageFactory;
+use Magento\Framework\Json\Helper\Data;
+use Psr\Log\LoggerInterface;
+use Salesfire\Salesfire\Helper\Feed\Generator;
+
+/**
+ * Salesfire Generate Controller
+ *
+ * @category   Salesfire
+ * @package    Salesfire_Salesfire
+ * @version.   1.3.0
+ */
+class Generate extends \Magento\Backend\App\Action
+{
+
+    protected $resultPageFactory;
+    protected $jsonHelper;
+
+    /**
+     * Constructor
+     *
+     * @param \Magento\Backend\App\Action\Context  $context
+     * @param \Magento\Framework\Json\Helper\Data $jsonHelper
+     */
+    public function __construct(
+        Context $context,
+        PageFactory $resultPageFactory,
+        Data $jsonHelper,
+        LoggerInterface $logger,
+        Generator $feedGenerator,
+    ) {
+        $this->resultPageFactory = $resultPageFactory;
+        $this->jsonHelper = $jsonHelper;
+        $this->logger = $logger;
+        $this->feedGenerator = $feedGenerator;
+        parent::__construct($context);
+    }
+
+    /**
+     * Execute view action
+     *
+     * @return \Magento\Framework\Controller\ResultInterface
+     */
+    public function execute()
+    {
+        $this->feedGenerator->execute();
+
+        try {
+            return $this->jsonResponse([
+                'status'    => 'complete',
+                'message'   => 'Your feeds have been generated.',
+            ]);
+        } catch (\Magento\Framework\Exception\LocalizedException $e) {
+            return $this->jsonResponse($e->getMessage());
+        } catch (\Exception $e) {
+            $this->logger->critical($e);
+            return $this->jsonResponse($e->getMessage());
+        }
+    }
+
+    /**
+     * Create json response
+     *
+     * @return \Magento\Framework\Controller\ResultInterface
+     */
+    public function jsonResponse($response = '')
+    {
+        return $this->getResponse()->representJson(
+            $this->jsonHelper->jsonEncode($response)
+        );
+    }
+}
